requirements.txt
fastapi
uvicorn
kafka-python
aiosmtplib
python-dotenv
pandas
openpyxl
python-docx
